Extract mutations logic into its own module
